renaming model to models
